// src/components/crypto/validation/enum.ts
export enum TaskType {
  SUBMISSION_PRIVATE = 'SUBMISSION_PRIVATE',
  SUBMISSION_HASH_ADDRESS = 'SUBMISSION_HASH_ADDRESS',
  SUBMISSION_IMAGE_ADDRESS = 'SUBMISSION_IMAGE_ADDRESS',
  SUBMISSION_IMAGE_ENTITY = 'SUBMISSION_IMAGE_ENTITY',
  SUBMISSION_ONLY_IMAGE = 'SUBMISSION_ONLY_IMAGE',
  SUBMISSION_ALL = ''
}